codex_artifacts/kits/custom-template-kit/requirements.txt
e2b==2.15.2
python-dotenv==1.2.2
